Modify get_peak_registration_hours method
The method was too complicated and hard to read. I changed it to simply
tally how many registries there were in each hour of the day instead of
also narrowing it down to the top three. This could be helpful to know
which hours to absolutely not run ads or something along those lines. I
also made the method outputs strings with ':00' appended to them for
higher readability.